@@ -6,24 +6,29 @@
   "main": "js/app.js",
   "private": true,
   "dependencies": {
-    "react": "^0.13.3",
-    "react-select": "^0.5.6",
-    "lodash": "^3.10.0"
+    "create-react-class": "^15.6.0",
+    "lodash": "^4.17.4",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
+    "react-router-dom": "^4.2.2",
+    "react-scripts": "^1.0.13",
+    "react-select": "^1.0.0-rc.10"
   },
   "devDependencies": {
-    "browserify": "^6.2.0",
-    "envify": "^3.0.0",
-    "reactify": "^0.15.2",
-    "watchify": "^2.1.1",
-    "babelify": "^6.1.3",
-    "node-sass": "^3.2.0",
-    "uglify-js": "^2.4.23"
+    "browserify": "^14.4.0",
+    "envify": "^4.1.0",
+    "reactify": "^1.1.1",
+    "watchify": "^3.9.0",
+    "babelify": "^7.3.0",
+    "node-sass": "^4.5.3",
+    "uglify-js": "^3.1.1"
   },
   "scripts": {
     "build-js": "browserify ./static/js/app.js -t [envify --NODE_ENV production] | uglifyjs -cm > static/dist/bundle.js",
-    "watch-js": "watchify static/js/app.js -d -o static/dist/bundle.js -v",
+    "watch-js": "watchify static/js/app.js -o static/dist/bundle.js",
     "watch-css": "node-sass -rw ./static/stylesheets/ -o ./static/dist",
-    "start": "npm run watch-css & npm run watch-js"
+    "start": "react-scripts start",
+    "build": "react-scripts build"
   },
   "author": "Scott Snider",
   "browserify": {
